automation-control-service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
redis==5.2.1
httpx==0.27.2
python-dotenv==1.0.1
py_eureka_client==0.11.2
pydantic==2.9.2